+++ src/proto_024_PtTALLiN/lib_plugin/plugin_registerer.ml
@@ -6,15 +6,6 @@
 (*                                                                           *)
 (*****************************************************************************)
 
-module Validation = struct
-  include Registerer.Registered
-
-  module Plugin = struct
-    include Plugin.Mempool
-    include Plugin.Block_validation
-  end
-end
-
 module RPC = struct
   module Proto = Registerer.Registered
   include Plugin.RPC
@@ -37,8 +28,6 @@ module Shell_helpers = struct
 
   let hash = Registerer.Registered.hash
 end
-
-let () = Protocol_plugin.register_validation_plugin (module Validation)
 
 let () = Protocol_plugin.register_rpc (module RPC)